% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flame.R
\name{geom_flame}
\alias{geom_flame}
\title{Create flame ploygons of Marine Heat Waves or Cold Spells.}
\usage{
geom_flame(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., stat.top = "int_cum", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat.top}{Choose which statistic to use to highlight the top event
found within the time series. Options are: \code{int_max} for the maximum
intensity of the event, \code{int_cum} for the cummulative intensity of
the event, \code{int_mean} for the mean intensity of the event and \code{dur}
for the duration of the event.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning. If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}
}
\value{
The function will create a ggplot2 geom indicating the climatology,
threshold and temperature, with the hot or cold events that meet the
specifications of Hobday et al. (2016) shown as flame polygons. The function
automagically detects if one is plotting MHWs or MCS and adjusts the colour
palette accordingly. These default colours may still be overridden as normal.
The top event detect during the selected time period, as determined by the
selected statistics given in \code{stat.top} will be visible in a
brighter colour. The default colour may be overidden in the aesthetics as well.
This function differs in use from \code{\link{event_line}}
in that it must be created as a ggplot() object. The benefit of this being
that one may add additional information to the figure as may be necessary.
}
\description{
Creates a ggplot2 geom of warm or cold events as per the second row of Figure 3 in
Hobday et al. (2016).
}
\examples{
t_dat <- make_whole(sst_WA)
res <- detect(t_dat, climatology_start = 1983, climatology_end = 2012) # using default values
mhw <- res$clim
mhw <- mhw[10580:10690,]

\dontrun{
library(ggplot2)
ggplot(mhw, aes(x = date, y = temp, thresh = thresh_clim_year, seas = seas_clim_year, event = event_no)) +
 geom_flame() +
 geom_text(aes(x = as.Date("2011-02-01"), y = 28, label = "Wow. Such heatwave. Many warm."))
}
}
\references{
Hobday, A.J. et al. (2016), A hierarchical approach to defining
marine heatwaves, Progress in Oceanography, 141, pp. 227-238,
doi: 10.1016/j.pocean.2015.12.014
}
\author{
Robert W. Schlegel
}
