% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotGeoms.R
\name{geom_lolli_plot}
\alias{geom_lolli_plot}
\title{Create a Timeline geom of Selected Event Metrics.}
\usage{
geom_lolli_plot(data, metric = "int_max", event_count = 3,
  xaxis = "date_start")
}
\arguments{
\item{data}{Output from the \code{\link{detect}} function.}

\item{metric}{One of \code{int_mean}, \code{int_max}, \code{int_cum} and \code{duration}.
Default is \code{int_cum}.}

\item{event_count}{The number of top events to highlight. Default is 3.}

\item{xaxis}{One of \code{event_no}, \code{date_start} or \code{date_peak}.
Default is \code{date_start}.}
}
\description{
Visualise a timeline of several event metrics as a 'lollipop' geom.
}
\examples{
t_dat <- make_whole(sst_NW_Atl)
res <- detect(t_dat, climatology_start = 1983, climatology_end = 2012) # using default values

\dontrun{
ggplot() + geom_lolli_plot(res, metric = "int_cum", event_count = 3, xaxis = "date_peak") +
geom_text(aes(x = as.Date("2007-01-01"), y = 375,
label = "One may clearly see\\njust how dramatic\\n these heatwaves are."))
}
}
\author{
Albertus J. Smit and Robert W. Schlegel
}

