% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWhole.R
\name{make_whole}
\alias{make_whole}
\title{Constructs a Continuous, Uninterrupted Time Series of Temperatures.}
\usage{
make_whole(data)
}
\arguments{
\item{data}{A data frame with columns headed \code{t} and \code{temp} for
date and temperature data, respectively. Daily data are expected, and
although missing values (NA) can be accommodated, the function is only
recommended when NAs occur infrequently, preferably at no more than 3
consequtive days.}
}
\value{
The function will return a data frame with three columns. The column
headed \code{doy} (day-of-year) is the Julian day running from 1 to 366, but
modified so that the day-of-year series for non-leap-years runs 1...59 and
then 61...366. For leap years the 60th day is February 29. See the example,
below. The \code{date} column is a series of dates of class \code{Date},
while \code{temp} is temperature. This time series will be uninterrupted and
continuous between the first and last dates of the input data.
}
\description{
Takes a series of dates and times, and if irregular (but ordered), inserts
missing dates and fills correpsonding temperatures with NAs.
}
\details{
Upon import, the package uses `zoo` and `lubridate` to process the input
date and temperature data. It reads in daily data with the time vector
specified as either \code{POSIXct} (e.g. "1982-01-01 02:00:00") or class
\code{date} (e.g. "1982-01-01"). The data may be an irregular time series,
but date must be ordered. The function constructs a complete time series
from the start date to the end date, and fills in the regions in the time
series where data are missing with NAs. Leap years are automatically
accommodated by 'zoo'.

This function can handle any amount of missing days, but this is not a
licence to actually use these data for the detection of anomalous thermal
events. Hobday et al. (2016) recommend gaps of no more than 3 days, which
may be adjusted by setting the \code{max_pad_length} argument of the
\code{\link{detect}} function. The longer and more frequent the gaps become
the lower the fidelity of the annual climatology and threshold that can be
calculated, which will not only have repercussions for the accuracy at which
the event metrics can be calculated, but also for the number of events that
can be detected.
}
\examples{
require(dplyr); require(reshape2); require(lubridate)
ts_dat <- make_whole(sst_WA)
clim_start <- "1983-01-01"
clim_end <- "2012-12-31"
ts_dat \%>\%
   filter(date >= clim_start & date <= clim_end) \%>\%
   dcast(doy ~ year(date), value.var = "temp", fill = as.numeric(NA)) \%>\%
   as.data.frame() \%>\%
   filter(doy >= 55 & doy <= 65)
}
\author{
Smit, A. J.
}

